Adding priority functionalites to ROS2.
>Summary
This change allows the user to use a new API "rclcpp::sched::SchedBase::set_sched_attr"
to set the (Linux) scheduling priority for each invocation of a subscriber/timer/servi-
ce/client.

Minor changes:
1. Using std::thread to initiate a new thread instead of pthread,
as it has better support for calling with multiple arguments and thus avoiding
creating a PthreadArg on heap, as in previous commits.
2. Have subscription/timer/client/service classes inherit SchedBase.
3. Some optimization and adjustments as suggested by @wagler

>Known Issues
sched::get_pid() is non-portable, as the pthread data structure is dependent on the
platform used and is not exposed to the user. Here we would not want the new thread
to set their own priority after being scheduled since the new thread may never be sche-
duled due to insufficient priority before setting its priority, but has de-facto
m-highest priority. We also did not choose to use the API provided by pthread (thus
avoiding the hack in sched_base.hpp:get_pid()) since glibc does not support
SCHED_DEADLINE-related attributes on Linux. See:
https://www.man7.org/linux/man-pages/man2/sched_setattr.2.html for detail.